Fix EZP-24367: Token error with remember me activated in Legacy and Symfony
> https://jira.ez.no/browse/EZP-24367

RememberMeToken doesn't accept to be manually set authenticated, which
is done by legacy mode. This is not needed as all
authentication/authorization process is fully done by legacy in that
case.

This PR removes the `setAuthenticated()` call on the token. Note that in
the profiler toolbar, username will appear in yellow (not
authenticated). This is perfectly safe as security is fully handled by
legacy kernel.
Further more, this PR also ensures that RememberMeListener is never
called when in legacy mode.