[Feature] Add command and composer command to setup legacy settings/designs symlinks
Without this feature, each and every user would have to find out the hard way that
Composer sometimes wipes out the ezpublish_legacy folder, and also that versioning
this folder is not really recommended.

The feature in this PR is primary aiming at upgrade/migration use cases, where
people are coming from 5.4 or 2014.x and similar and need to move over settings and designs.

Review wanted from partners that have their own way of doing this, as you will know
better what you need.

*Why no extension folder handling?*
Extensions are handled by composer and also by legacy bundles, so if you need an extension in
src folder you can make app bundle an legacy bundle.
Hence it felt wrong to give another option for this which could be confusing.

*Why no storage and storage/packages folder handling?*
As site storage folder is not typically versioned, I guess this should be handled by something else.
Input on how people tend to deal with this more then welcome.