Throw exception when module or view is not found.
The legacy ezpublish does not return the 404 response code for any
request to a page that does not exist. It simply redirects all invalid
requests to the login page with a resonse code 200.

To fix this issue, method `ezpkernelweb::dispatchLoop()` has been
changed to throw exception when the requested module or view does not
exist. In order to redirect the invalid requests to a 404 error page,
the exception thrown by the legacy ezpublish needs to be converted to
`HttpNotFoundException`.

Refer to ezsystems/ezpublish-legacy#1379 for details.

| Question           | Answer
| ------------------ | ------------------
| **JIRA issue**     | n/a
| **Bug**            | yes
| **New feature**    | no
| **Target version** | `v2.0.0`
| **BC breaks**      | yes
| **Tests pass**     | yes
| **Doc needed**     | no

**TODO**:
- [ ] Implement feature / fix a bug.
- [ ] Implement tests.
- [ ] Fix new code according to Coding Standards (`$ composer fix-cs`).
- [x] Ask for Code Review.